+++ echos_lab/engines/post_maker.py
@@ -13,7 +13,7 @@ from echos_lab.crypto_lib import crypto_connector
 from echos_lab.db import models
 from echos_lab.engines import agent_context, prompts
 from echos_lab.engines import legacy
-from echos_lab.engines.personalities.profiles import AgentProfile, LegacyAgentProfile
+from echos_lab.engines.profiles import AgentProfile, LegacyAgentProfile
 from echos_lab.engines.prompts import (
     SubTweetEvaluation,
     TweetEvaluation,
