+++ echos_lab/engines/prompts.py
@@ -12,7 +12,7 @@ from echos_lab.common import utils
 from echos_lab.crypto_lib import crypto_connector
 from echos_lab.db import models
 from echos_lab.engines.agent_context import AgentContext
-from echos_lab.engines.personalities.profiles import AgentProfile, LegacyAgentProfile
+from echos_lab.engines.profiles import AgentProfile, LegacyAgentProfile
 from echos_lab.twitter.types import MentionType, TweetMention
 
 base_path = Path(__file__).parent
