+++ echos_lab/main.py
@@ -10,8 +10,8 @@ from echos_lab.common.logger import logger
 from echos_lab.crypto_lib import crypto_connector
 from echos_lab.db import db_connector, db_setup
 from echos_lab.engines import full_agent, images, post_maker
-from echos_lab.engines.personalities import profiles
-from echos_lab.engines.personalities.profiles import AgentProfile, LegacyAgentProfile
+from echos_lab.engines import profiles
+from echos_lab.engines.profiles import AgentProfile, LegacyAgentProfile
 from echos_lab.slack.client import SlackClient
 from echos_lab.telegram import telegram_client
 from echos_lab.twitter import (
