+++ echos_lab/twitter/twitter_workflows.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import Session
 from echos_lab.common.logger import logger
 from echos_lab.common.utils import with_db
 from echos_lab.engines import post_maker, prompts
-from echos_lab.engines.personalities.profiles import AgentProfile
+from echos_lab.engines.profiles import AgentProfile
 from echos_lab.twitter import twitter_client, twitter_pipeline, twitter_poster
 from echos_lab.twitter.types import (
     MEME_RATING_THRESHOLD,
