+++ tests/integration/test_db_migration.py
@@ -7,7 +7,7 @@ from sqlalchemy.orm import Session
 
 from echos_lab.db.migrations import json_to_postgres
 from echos_lab.db.models import QueryType, TwitterQueryCheckpoint, TwitterUser
-from echos_lab.engines.personalities.profiles import AgentProfile
+from echos_lab.engines.profiles import AgentProfile
 from echos_lab.twitter import twitter_client
 
 # Path constants
